@@ -25,11 +25,10 @@ ext {
 }
 
 android {
-    compileSdkVersion 28
-
+    compileSdkVersion 30
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 28
+        targetSdkVersion 30
         versionCode 1
         versionName "3.2.0-beta.1"
     }
@@ -37,7 +36,7 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:support-compat:28.0.0'
+    implementation 'androidx.core:core:1.5.0'
     api 'com.github.barteksc:pdfium-android:1.9.0'
 }
 
