@@ -1,23 +1,24 @@
 buildscript {
     repositories {
         google()
+        mavenCentral()
         jcenter()
     }
 }
 
 repositories {
     google()
+    mavenCentral()
     jcenter()
 }
 
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
-
+    compileSdkVersion 30
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 28
+        targetSdkVersion 30
         versionCode 3
         versionName "3.0.0"
     }
@@ -26,7 +27,7 @@ android {
 
 dependencies {
     implementation project(':android-pdf-viewer')
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'org.androidannotations:androidannotations-api:4.6.0'
-    annotationProcessor "org.androidannotations:androidannotations:4.6.0"
+    implementation 'androidx.appcompat:appcompat:1.3.0'
+    implementation 'org.androidannotations:androidannotations-api:+'
+    annotationProcessor "org.androidannotations:androidannotations:+"
 }
